+++ Sources/SwiftiumTestingKit/SwiftiumTestingKit.swift
@@ -1,0 +1,3 @@
+struct SwiftiumTestingKit {
+    var text = "Hello, World!"
+}