+++ tag-release.sh
@@ -6,18 +6,14 @@ if [ $# -lt 1 ] ; then
 fi
 
 VERSION=$1
-TAG=v$VERSION
+TAG=$VERSION
 
 echo "Have you updated the changelog for version $VERSION ? (ctrl-c to go update it)"
 read
 
 set -e
 
-carthage build --no-skip-current && carthage archive SwiftiumTestingKit
-
 echo "Creating tag $TAG and pushing it to github"
 git tag $TAG
 git push --tags
-
-echo "You can now upload SwiftiumTestingKit.framework.zip and edit release notes from https://github.com/openium/SwiftiumTestingKit/releases/edit/$TAG"