+++ Sources/SwiftiumTestingKit/SwiftiumTestingKit.swift
@@ -1,3 +1,0 @@
-struct SwiftiumTestingKit {
-    var text = "Hello, World!"
-}