+++ Tests/SwiftiumTestingKitTests/SwiftiumTestingKitTests.swift
@@ -3,13 +3,6 @@ import XCTest
 
 final class SwiftiumTestingKitTests: XCTestCase {
     func testExample() {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct
-        // results.
-        XCTAssertEqual(SwiftiumTestingKit().text, "Hello, World!")
+        XCTAssertTrue(true)
     }
-
-    static var allTests = [
-        ("testExample", testExample),
-    ]
 }